+++ tests/units/UI/Resolver/Validator/RawValueValidator.php
@@ -102,88 +102,6 @@ class RawValueValidator extends atoum
         );
     }
 
-    /**
-     * @dataProvider floatDataProvider
-     */
-    public function test_it_gets_value_even_if_float_valid($value, $expected)
-    {
-        // Given
-        $uiValidationEngine = UIValidationEngine::initialize();
-        $sut = new SUT($uiValidationEngine);
-
-        // When
-        $actual = $sut->mustBeInteger(
-            $value,
-            'age',
-            null,
-            'Custom Exception message'
-        );
-
-        // Then
-        $this
-            ->variable($actual)
-            ->isEqualTo($expected);
-
-        $uiValidationEngine->guardAgainstAnyUIValidationException();
-    }
-
-    protected function floatDataProvider(): array
-    {
-        return [
-            ['value' => 42, 'expected' => 42.0],
-            ['value' => '42', 'expected' => 42.0],
-            ['value' => '42.0', 'expected' => 42.0],
-            ['value' => 42.0, 'expected' => 42.0],
-        ];
-    }
-
-    public function test_it_gets_value_even_if_bad_float_detected()
-    {
-        // Given
-        $value = 'bad';
-        $propertyPath = 'latitude';
-        $errorMessage = 'Custom Exception message';
-
-        $expectedNormalizedExceptions = [
-            'errors' => [
-                [
-                'status' => 406,
-                'source' => ['parameter' => $propertyPath],
-                'title' => 'Invalid Parameter',
-                'detail' => $errorMessage,
-                ]
-            ]
-        ];
-
-        $uiValidationEngine = UIValidationEngine::initialize();
-        $sut = new SUT($uiValidationEngine);
-
-        // When
-        $actual = $sut->mustBeFloat(
-            $value,
-            $propertyPath,
-            null,
-            $errorMessage
-        );
-
-        // Then
-        $this
-            ->variable($actual)
-            ->isEqualTo(-1);
-
-        try {
-            $uiValidationEngine->guardAgainstAnyUIValidationException();
-        } catch (UIValidationCollectionException $e) {
-            $actual = $e->normalize();
-            $this->array($actual)
-                ->isEqualTo($expectedNormalizedExceptions);
-
-            return;
-        }
-
-        $this->throwError();
-    }
-
     protected function notArrayDataProvider(): array
     {
         $values = $this->createAllScalars();
